feat(118850): altera campo total_embalagens para FloatField

+++ sme_terceirizadas/pre_recebimento/models/cronograma.py
@@ -125,9 +125,7 @@ class EtapasDoCronograma(ModeloBase):
     parte = models.CharField(blank=True, max_length=15)
     data_programada = models.DateField("Data Programada", blank=True, null=True)
     quantidade = models.FloatField(blank=True, null=True)
-    total_embalagens = models.PositiveIntegerField(
-        "Total de Embalagens", blank=True, null=True
-    )
+    total_embalagens = models.FloatField("Total de Embalagens", blank=True, null=True)
 
     def __str__(self):
         if self.etapa and self.cronograma:
